fuzdl::driver
Function with_driver

pub async fn with_driver<Out, Fut: Future<Output = Out> + Send>(
    f: impl FnOnce(WebDriver) -> Fut + Send + 'static,
) -> Result<Option<Out>>

Create a WebDriver and run the callback function with the driver as the argument.

Example

driver::with_driver(|driver| async move {
    driver.get("https://www.google.com").await.unwrap();
    assert_eq!(driver.title().await.unwrap(), "Google");
})
.await
.unwrap();

Return Value

The outer Result is of the Err variant if an error occurs in the driver creation or cleanup steps. The inner Option is of the None variant if a kill signal is received before the callback function completes. In both cases, the driver resources are properly cleaned up before returning.

Resource Leaks

In some cases, the resources may not be properly cleaned up. These cases include but are not limited to:

The callback function panics.

The driver creation functions panics.

When can we have support for async drop in Rust?